[infra-proxy-service] Add TestHealthGRPC & TestVersion test cases.  - Add missing Health API to the GRPC test server.  - Move code into test helper.
